Add files via upload
Assig1:Write a program that reads a positive integer and checks if it is a perfect square.
________________________________________________________________________________________________________
Assig2:Write a program to make a simple calculator using switch-case. The calculator takes the operation
(+ or – or * or /) and takes the two input arguments and
print the results.
________________________________________________________________________________________________________
Assig3:Write a program that takes one character and checks if it alphabet or not.

Assig4:Write program that converts any letter from lowercase to uppercase.
________________________________________________________________________________________________________
Assig5:Write a program that reads two integers and checks if the first is multiple of the second.
________________________________________________________________________________________________________
Assig6: Write a program that calculates the required heater activation time according to the input temperature of water.
- if the input temperature is from 0 to 30, then required
heating time = 7 mins.
- if the input temperature is from 30 to 60, then required
heating time = 5 mins.
- if the input temperature is from 60 to 90, then required
heating time = 3 mins.
- if the input temperature is more than 90, then required
heating time = 1 mins.
- if temperature is invalid, display "Invalid input"
Example:
Input = 10 -> output = 7
Input = 35 -> output = 5
________________________________________________________________________________________________________
Assig7: Write a program to add two floating numbers. Determine the integer floor of the sum. The floor is the truncated float value, anything after the decimal point is dropped.